# skills/kanban-openapi.yaml
openapi: 3.0.0
info:
  title: AI Marketing Crew Agent API
  description: API for AI Agents to interact with the AI Marketing Crew collaboration board. Enables task creation, state transition, human-in-the-loop requests, and task retrieval.
  version: 1.0.0
servers:
  - url: https://amc-kanban.immedi.ai/api
    description: Production Server
  - url: http://localhost:3000/api
    description: Local Development Server
paths:
...
  /brands/{brandId}/drafts:
    get:
      summary: List brand content drafts
      operationId: listBrandDrafts
      parameters:
        - name: brandId
          in: path
          required: true
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: q
          in: query
          required: false
          schema:
            type: string
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Draft list
    post:
      summary: Create content draft
      operationId: createBrandDraft
      description: Create a saved draft. Call submit endpoint when ready.
      parameters:
        - name: brandId
          in: path
          required: true
          schema:
            type: string
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DraftInput'
      responses:
        '201':
          description: Draft created
  /brands/{brandId}/drafts/{draftId}:
    get:
      summary: Read one content draft
      operationId: getBrandDraft
      parameters:
        - name: brandId
          in: path
          required: true
          schema:
            type: string
        - name: draftId
          in: path
          required: true
          schema:
            type: string
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Draft details
    patch:
      summary: Update content draft
      operationId: updateBrandDraft
      parameters:
        - name: brandId
          in: path
          required: true
          schema:
            type: string
        - name: draftId
          in: path
          required: true
          schema:
            type: string
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DraftInput'
      responses:
        '200':
          description: Draft updated
  /brands/{brandId}/drafts/{draftId}/submit:
    patch:
      summary: Submit draft for approval or direct delivery
      operationId: submitBrandDraft
      description: Auto-pilot brands publish/schedule directly. Boss-approval brands create pending review ActionItem.
      parameters:
        - name: brandId
          in: path
          required: true
          schema:
            type: string
        - name: draftId
          in: path
          required: true
          schema:
            type: string
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                note:
                  type: string
      responses:
        '200':
          description: Draft submitted
...
  /brands/{brandId}/topics:
    get:
      summary: List TopicFeed research documents for a brand
      operationId: listBrandTopicFeeds
      description: Read markdown research topics, trend observations, competitor notes, and content angles for a brand. AI agents can use this before drafting content.
      parameters:
        - name: brandId
          in: path
          required: true
          schema:
            type: string
        - name: q
          in: query
          required: false
          schema:
            type: string
        - name: tag
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [active, archived, all]
        - name: Authorization
          in: header
          required: true
          description: Bearer token received from your Initialization Instruction
          schema:
            type: string
      responses:
        '200':
          description: TopicFeed list
          content:
            application/json:
              schema:
                type: object
                properties:
                  topics:
                    type: array
                    items:
                      $ref: '#/components/schemas/TopicFeed'
    post:
      summary: Create a TopicFeed markdown research document
      operationId: createBrandTopicFeed
      description: Write a markdown research topic for the brand. Use this for research findings, content angles, trend notes, and competitor observations.
      parameters:
        - name: brandId
          in: path
          required: true
          schema:
            type: string
        - name: Authorization
          in: header
          required: true
          description: Bearer token received from your Initialization Instruction
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - markdown
              properties:
                title:
                  type: string
                summary:
                  type: string
                  nullable: true
                tags:
                  type: array
                  items:
                    type: string
                sourceUrl:
                  type: string
                  nullable: true
                markdown:
                  type: string
                  description: Markdown document body.
      responses:
        '201':
          description: TopicFeed created
...
components:
  schemas:
...
    TopicFeed:
      type: object
      properties:
        id:
          type: string
        brandId:
          type: string
        title:
          type: string
        markdown:
          type: string
        summary:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        sourceUrl:
          type: string
          nullable: true
        status:
          type: string
        createdById:
          type: string
          nullable: true
        createdByType:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    DraftInput:
      type: object
      required:
        - caption
      properties:
        caption:
          type: string
        captionLang:
          type: string
          default: en
        accountId:
          type: string
          nullable: true
        mediaUrls:
          type: array
          items:
            type: string
        assetIds:
          type: array
          items:
            type: string
        hashtags:
          type: array
          items:
            type: string
        scheduledAt:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
          description: Usually draft. Use submit endpoint rather than writing pending_review directly.
        agentNote:
          type: string
...